Split out the wifi root node device, since we may want to support no-wifi.
